Adequacoes e novas funcionalidades.

+++ JPBM/Entidades/TipoContato.cs
@@ -2,7 +2,7 @@
 {
     public class TipoContato
     {
-        public int TipoContatoId { get; set; }
+        public byte TipoContatoId { get; set; }
         public string Descricao { get; set; }
     }
 }
